64 bit fixes for arinc data. Noticed the problem in tests/sync_server_dump. In Process::get/setEffectiveCapability: do cap_free on exceptions in     cap_get/set_flag, cap_set_proc to prevent possible small memory leak.

+++ src/nidas/dynld/raf/DSMArincSensor.h
@@ -109,7 +109,7 @@ namespace nidas { namespace dynld { namespace raf {
     void printStatus(std::ostream& ostr) throw();
 
     /** This contains a switch case for processing all labels. */
-    virtual float processLabel(const long data) = 0;
+    virtual float processLabel(const int data) = 0;
 
     /** Extract the ARINC configuration elements from the XML header. */
     /// example XML:
@@ -139,7 +139,7 @@ namespace nidas { namespace dynld { namespace raf {
     bool sim_xmit;
   };
 
-  typedef SampleT<unsigned long> ArincSample;
+  typedef SampleT<unsigned int> ArincSample;
 
 }}}	// namespace nidas namespace dynld namespace raf
